S3.3b.1: full generic class method monomorphization
Generic class instances are typed `Type::Generic{class_id, args}` end-to-end
and methods using `T` are monomorphized per concrete type at every call site,
restoring the §3.3 ARCHITECTURE_REFACTOR.md non-negotiable: no `Type::Var`
remains in any non-template function by codegen time. Unblocks the strict
`assert_no_var_remaining` invariant on all locals (no skipping).

- Move `derive_subst` from `optimizer::monomorphize::derive` to
  `pyaot_types::derive`. Single source of truth for both free-function
  monomorphize and class-constructor result-type derivation; optimizer's
  duplicate copy deleted.
- Pack `*args` into a runtime tuple in `lower_class_method_call` at every
  emit site (CallVirtual, CallVirtualNamed, fallback CallDirect, dunder,
  classmethod, staticmethod). Mirror of `Lowering::resolve_call_args`'s
  pattern for free calls; prerequisite for safe pre-mono devirt.
- `lower_class_instantiation` emits `Type::Generic{class_id, [args]}`
  directly when the class has `type_params`: pre-resolves `__init__` user
  args, derives `subst` against the pre-stamped `__init__` signature,
  substitutes `type_params`. Removes the post-hoc alloc-result type
  promotion in `lower_assign` (was an ad-hoc escape hatch per §5).
- Re-enable `DevirtualizePass` between WPA pass 1 and `monomorphize::run`.
  After var-arg packing this is safe: `CallVirtual` already matches the
  callee ABI, so post-devirt `CallDirect` does too.
- Monomorphize fixpoint: after specializing a CallDirect, update the dest
  local's type to the specialization's return type, propagate forward
  through `Copy { dest, src: Local }` aliases, re-enqueue the caller.
  Closes chained method calls like `b.rebox().get()`.
- Restore strict `assert_no_var_remaining` on locals (was relaxed; now
  required after the fixpoint loop concretizes all caller-local types).
- T-everywhere test fixture: `IntWrapper(Generic[T])` with `val: T`,
  `__init__(self, v: T)`, `unwrap(self) -> T`, `set_value(self, v: T)`,
  `transform(self, v: T) -> T`. PEP-695 `TickBox[K]` with `tag(self,
  label: K) -> K`, `get_label(self) -> K`. Plus `Box[T]` chained-spec
  test (`b.rebox().get()`), `Counter` + `@runtime_checkable Protocol
  Unwrappable` polymorphic-receiver test.
- Unit tests: `pyaot_types::derive_subst` (Bool/Int distinct guard for
  numeric tower), `devirtualize` Generic-receiver + builtin-list-no-vtable,
  `dce::reachability` phi-source pruning lock-in, `constfold::propagate`
  Branch→Goto phi-source pruning lock-in.
- Drive-by: fix two clippy `unnecessary_cast` errors in codegen-cranelift
  (unblocks `clippy --release -D warnings` on the workspace).

Acceptance: `cargo build/fmt/clippy/test --workspace --release` green.
29 method specializations in test_generics MIR, 1 surviving CallVirtual
(Protocol receiver), byte-for-byte CPython parity. Binary size +0.25%
(6 552 704 → 6 569 296), well under §3.7's +20% gate.

Inheritance widening (`class C(Base[T])`) is the separate S3.3c task —
requires WPA/devirt subclass-set analysis.

+++ crates/codegen-cranelift/src/runtime_calls/mod.rs
@@ -69,7 +69,7 @@ pub fn compile_runtime_call(
             if std::ptr::eq(*def, &RT_INSTANCE_GET_FIELD_F64) {
                 let inst_val = load_operand(builder, &args[0], ctx.symbols.var_map);
                 if let Operand::Constant(mir::Constant::Int(off)) = &args[1] {
-                    let byte_off = INSTANCE_FIELDS_OFFSET as i32 + (*off as i32) * PTR_SIZE as i32;
+                    let byte_off = INSTANCE_FIELDS_OFFSET + (*off as i32) * PTR_SIZE as i32;
                     let result =
                         builder
                             .ins()
@@ -83,7 +83,7 @@ pub fn compile_runtime_call(
             if std::ptr::eq(*def, &RT_INSTANCE_SET_FIELD_F64) {
                 let inst_val = load_operand(builder, &args[0], ctx.symbols.var_map);
                 if let Operand::Constant(mir::Constant::Int(off)) = &args[1] {
-                    let byte_off = INSTANCE_FIELDS_OFFSET as i32 + (*off as i32) * PTR_SIZE as i32;
+                    let byte_off = INSTANCE_FIELDS_OFFSET + (*off as i32) * PTR_SIZE as i32;
                     let value_val =
                         load_operand_as(builder, &args[2], ctx.symbols.var_map, cltypes::F64);
                     builder
